Finally prove some old theorems about `foldBag` in GHC
In particular, I prove that `foldBag` is `referenceFoldBag` iff either
(1) the `Bag` is well-formed, `f` is associative, and `z` is a right identity,
    or
(2) `f` is associative, and `z` is a (two-sided) identity.

This confirms my long-standing, largely immaterial, suspicion that you don't
need the `left_id` part, while also demonstrating why I needed it in my proofs.

I don't prove a big "iff" theorem; I prove equalities conditioned on either (1)
or (2), and then provide counterexamples for the other cases.